smsServer/flaskr/requirements.txt
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pymongo==4.6.3
python-dateutil==2.9.0.post0
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
Werkzeug==3.0.1